pyproject.toml
[tool.poetry]
name = "hueplanner"
version = "0.3.0"
description = ""
authors = ["Alex Zakharov <dokzlo13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.1"
pydantic = {extras = ["dotenv"], version = "^2.5.3"}
astral = "^3.2"
geopy = "^2.4.1"
timezonefinder = "^6.2.0"
environs = "^11.0.0"
structlog = "^24.4.0"
uvloop = "^0.21.0"
tzdata = "^2024.1"
aiosqlite = "^0.20.0"
pyyaml = "^6.0.1"
pydantic-settings = "^2.5.2"
tzlocal = "^5.2"
jq = "^1.8.0"
pyaml-env = "^1.2.1"
pytimeparse2 = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.25.0"
black = "^24.8.0"
isort = "^5.10.1"
bandit = "^1.7.4"
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
flake8-isort = "^6.1.1"
flake8-bandit = "^4.1.1"
mypy = "^1.11.2"

[tool.bandit]
targets = ["hueplanner", "tests"]

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pip-cache
  | \.poetry
  | \.venv
  | \.vscode
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
